% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Add_geom}
\alias{Add_geom}
\title{Add geometries and calculated distance}
\usage{
Add_geom(con, save_dir)
}
\arguments{
\item{con}{Connection to db}

\item{save_dir}{Dirrectory where to save geometries}
}
\description{
Add geometries to lucas harmonized table:
- location of theoretical point(th_geom) from fields th_long, the_lat
- location of lucas survey (gps_geom) from fields gps_long, gps_lat
- lucas transect geometr (trans_geom) from fields gps_long, gps_lat
- distance between theoretical and survey point (th_gps_dist)
}
\examples{
\dontrun{
Add_geom(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
