% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Add_new_cols}
\alias{Add_new_cols}
\title{Add new columns to tables}
\usage{
Add_new_cols(con, years)
}
\arguments{
\item{con}{Connection to db}

\item{years}{Numeric vector of years to be harmonised}
}
\description{
Adds new columns to all table that will be necessary for when tables are merged. Includes
letter group - first level of LUCAS land cover/land use classification system
year - year of survey
file_path_gisco_n/s/e/w/p - file path to full HD images on ESTAT GISCO cloud service for North, South, East, West, and Point images
}
\examples{
\dontrun{
Add_new_cols(con, c(2006, 2009, 2012, 2015, 2018))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
