% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Add_revisit}
\alias{Add_revisit}
\title{Add revisit column}
\usage{
Add_revisit(con)
}
\arguments{
\item{con}{Connection to db}
}
\description{
Adds revisit column to lucas harmonized table to show the number of times between the years when the point was revisited.
}
\examples{
\dontrun{
Add_revisit(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
