% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Align_Map_CSVs}
\alias{Align_Map_CSVs}
\title{Align mapping CSVs}
\usage{
Align_Map_CSVs(mapp_csv_folder, years)
}
\arguments{
\item{mapp_csv_folder}{Directory where mapping files are stored}

\item{years}{Numeric vector of years to be harmonised}
}
\description{
Corrects any typo, spelling mistake, or spelling difference in the user-created mapping CSVs, used to generate labels in subsequent User_friendly() function by alligning them to the mapping CSV of the latest survey.
}
\examples{
\dontrun{
Align_Map_CSVs('/data/LUCAS_harmo/data/mappings', c(2006, 2009, 2012, 2015, 2018))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
