% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Assert_files}
\alias{Assert_files}
\title{Assert files}
\usage{
Assert_files(data_dir)
}
\arguments{
\item{data_dir}{Character. Folder where you saved all the micro data downloaded from EUROSTAT}
}
\value{
Nothing if OK error if failed
}
\description{
Check that the user has downoad all the files needed
}
