% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_lmd.R
\name{bind_lmd}
\alias{bind_lmd}
\title{Combine LMD from different countries in one SpatialPointsDataFrame}
\usage{
bind_lmd(lmddir, countries, year = 2012, cols = NULL,
  dst = file.path(lmddir, paste0("LMD", year, ".shp")),
  dst_overwrite = FALSE, ...)
}
\arguments{
\item{lmddir}{directory with the LMD}

\item{countries}{countries to be merged}

\item{year}{year of LMD}

\item{cols}{columns of LMD to be kept}

\item{dst}{filename of the merged data (ESRI shaefile)}

\item{dst_overwrite}{overwrite the destination file if it exists?}
}
\value{
\code{\link{SpatialPointsDataFrame}}
}
\description{
Combine LMD from different countries in one SpatialPointsDataFrame
}

