% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Check_Map_CSVs}
\alias{Check_Map_CSVs}
\title{Check mapping CSVs}
\usage{
Check_Map_CSVs(mapp_csv_folder, years, save_dir)
}
\arguments{
\item{mapp_csv_folder}{Directory where mapping files are stored}

\item{years}{Numeric vector of years to be harmonised}

\item{save_dir}{Directory where new C3 legends will be saved. Idealy (for consistency's sake) this directory should be the support_dir from main.R script.}
}
\description{
Consistency check for Allign_map_CSVs function and creation of a new document with the explicit legends documented in document C3 (Classification) from LUCAS surveys.
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
