% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_transect.R
\name{colnames_transect}
\alias{colnames_transect}
\title{Get column names of the transect attributes}
\usage{
colnames_transect(x)
}
\arguments{
\item{x}{a LMD data frame}
}
\value{
character vector with the transect names
}
\description{
Get column names of the transect attributes
}

