% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Connect_to_db}
\alias{Connect_to_db}
\title{Conntect to DB}
\usage{
Connect_to_db(user, host, port, password, dbname)
}
\arguments{
\item{user}{Character. User of the database}

\item{host}{Character. Host of the DB}

\item{port}{Integer. Port to connect to usually 5432}

\item{password}{Character. Password to access to the DB}

\item{dbname}{Character}
}
\value{
conection to the db
}
\description{
connect to the db where you want to upload all LUCAS points
}
\examples{
\dontrun{
con <- Connect_to_db("andrrap", "localhost", 5432,"andrrap","andrrap")
con <- Connect_to_db("martlur", "/var/run/postgresql", 5432,"martlur","postgres")
con <- Connect_to_db("postgres", "172.15.0.10", 5432,"test","postgres")}
}
