% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Consistency_check}
\alias{Consistency_check}
\title{Consistency checks}
\usage{
Consistency_check(con, years, manChangedVars)
}
\arguments{
\item{con}{Connection to db}

\item{years}{Numeric vector of years to be harmonised}

\item{manChangedVars}{File path to csv of attributes and relevant years to which manual manipulation has been done and thus cannot clear a consistency of values check}
}
\description{
Perform consistency checks on newly created tables to ensure conformity in terms of column order and data types
}
\examples{
\dontrun{
Consistency_check(con, c(2006, 2009, 2012, 2015, 2018))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
