% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Correct_long}
\alias{Correct_long}
\title{Harmonize long values in all tables}
\usage{
Correct_long(lucas2009)
}
\arguments{
\item{lucas2009}{Dataframe with the 2009 data}
}
\description{
Harmonize long values in all tables
}
\section{Functions}{
\itemize{
\item \code{Correct_long}: the long values of 2009 data by applying th_ew to th_long and erase this column
}}

\examples{
\dontrun{
Correct_long(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
