% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Correct_th_loc}
\alias{Correct_th_loc}
\title{Correct theoretical long lat}
\usage{
Correct_th_loc(con, lucas_grid)
}
\arguments{
\item{con}{Connection to db}

\item{lucas_grid}{File path to the csv file of the latest LUCAS grid}
}
\description{
Applying a correction of the values of columns th_long and th_lat according to the latest LUCAS grid
}
\examples{
\dontrun{
Correct_th_loc(con, lucas_grid)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
