% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_lmd.R
\name{create_spatial_lmd}
\alias{create_spatial_lmd}
\title{Convert LMD data to a SpatialPointsDataFrame}
\usage{
create_spatial_lmd(df, shp = NULL, ...)
}
\arguments{
\item{df}{data frame as returned by \code{\link{get_lmd}}}

\item{shp}{filename in case the spatial points data frame should be stored to a shapefile, \code{NULL} otherwise.}

\item{...}{passed to \code{\link{writeOGR}}}
}
\value{
\code{\link{SpatialPointsDataFrame}}
}
\description{
Create a \code{\link{SpatialPointsDataFrame}} from a LMD (-like) data frame as returned by \code{\link{get_lmd}}.
}

