% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Create_tags}
\alias{Create_tags}
\title{Create tags for harmonized table}
\usage{
Create_tags(con)
}
\arguments{
\item{con}{Connection to db}
}
\description{
Create database tags (primary key), index, and spatial index and a new id column for the harmonized table
}
\examples{
\dontrun{
Create_tags(con)}de
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
