% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all.R
\name{download_all}
\alias{download_all}
\title{Download all countries of all years and create shapefiles.}
\usage{
download_all(lmdddir, years = c(2006, 2009, 2012, 2015), countries = NULL,
  create_single_country_shape = TRUE, create_all_countries_shape = TRUE)
}
\arguments{
\item{lmdddir}{The base directory for the data. In this, a directory
will be ceated for each year.}

\item{years}{The years for which to download the data. Possible: 2006, 2009, 2012, 2015.}

\item{countries}{The countries to be downloaded in LMD code style 
(see \code{\link{(LMD code style)}}\code{$code}).
If \code{NULL} all available countries of a year are downloaded.}

\item{create_single_country_shape}{Should shapefiles be created for each country?}

\item{create_all_countries_shape}{Should a shapefile be created containing the 
points of all countries?}
}
\description{
Download all countries of all years and create shapefiles.
}

