% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_lmd.R
\name{download_lmd}
\alias{download_lmd}
\title{Download a LMD file - but better use get_lmd.}
\usage{
download_lmd(code, year, destfile = NULL, keep_original = FALSE,
  return_df = FALSE)
}
\arguments{
\item{code}{Country code (LUCAS code)}

\item{year}{Year}

\item{destfile}{File to write the file to}

\item{keep_original}{Keep the orignal downloaded file?}

\item{return_df}{Return the data frame? Else NULL is returned.}
}
\value{
data frame or NULL
}
\description{
Download a LMD file - but better use get_lmd.
}

