% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Final_order_cols}
\alias{Final_order_cols}
\title{Final column order}
\usage{
Final_order_cols(con)
}
\arguments{
\item{con}{Connection to db}
}
\description{
Re-order columns of final tables
}
\examples{
\dontrun{
Final_order_cols(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
