% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_country_code.R
\name{get_country_code}
\alias{get_country_code}
\title{Convert country codes}
\usage{
get_country_code(x = NULL, code_type = "lmd")
}
\arguments{
\item{x}{list of country names or country codes (lmd, gadm types). 
See \code{\link{get_lmd_overview}}. If \code{NULL} all available codes of 
type \code{code_type} are returned.}

\item{code_type}{the code type to be returned}
}
\value{
character vector with codes
}
\description{
Convert country codes
}

