% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename.R
\name{get_filename}
\alias{get_filename}
\title{Get default filenames used to store downloaded LMD data}
\usage{
get_filename(lmddir, year, country, type = "xls")
}
\arguments{
\item{lmddir}{base directory}

\item{year}{the year of the LMD data}

\item{country}{the country code}

\item{type}{the data format, xls or shp}
}
\value{
the filename as character
}
\description{
Get default filenames used to store downloaded LMD data
}

