% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm.R
\name{get_gadm}
\alias{get_gadm}
\title{Download Global Administrative Areas (GADM)}
\usage{
get_gadm(A3code, level = 0, destdir = NULL, filename_shp = NULL,
  overwrite_shp = FALSE, ...)
}
\arguments{
\item{A3code}{Three letter ISO 3166 code(s) (see e.g. \url{http://www.davros.org/misc/iso3166.html})}

\item{level}{administrative level}

\item{destdir}{destination directory to store the data}

\item{...}{passed to \code{\link{writeOGR}}}

\item{saveShp}{if not \code{NULL} the data is saved as  ESRI Shapefile. If \code{TRUE} or \code{'defualt'} it is written to a default filename and loction else it is assumed that the filename is given.}
}
\value{
\code{\link{SpatialPointsDataFrame}}
}
\description{
Download Global Administrative Areas (GADM)
}
\seealso{
\url{http://www.gadm.org/}
}

