% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LC1xLU1_counts.R
\name{get_LC1xLU1_counts}
\alias{get_LC1xLU1_counts}
\title{Get frequency table of LC1 & LU1 occurences}
\usage{
get_LC1xLU1_counts(df, as_df = TRUE)
}
\arguments{
\item{df}{a data fram containing the columns LC1 and LU1}

\item{as_df}{\code{TRUE} retuns the result as a data frame with three columns: LC1, LU1 and count. Else a table.}
}
\value{
according to \code{as_df} a data frame or a table
}
\description{
Get frequency table of LC1 & LU1 occurences
}

