% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmd.R
\name{get_lmd}
\alias{get_lmd}
\title{Download and/or load LUCAS Micro Data (LMD)}
\usage{
get_lmd(destdir = NULL, countries = NULL, years = NULL, cols = NULL,
  overwrite = FALSE, spatial = FALSE, save_shp = FALSE,
  keep_original = FALSE, verbosity = 1)
}
\arguments{
\item{destdir}{destination directory to store the data}

\item{countries}{the countries to be downloaded}

\item{years}{the years to be downloaded}

\item{cols}{the columns (attributes) to keep}

\item{overwrite}{should existing files be overwritten}

\item{spatial}{if \code{TRUE} a \code{\link{SpatialPointsDataFrame}} is returned}

\item{save_shp}{if \code{TRUE} the data is tored as ESRI shapefile}

\item{keep_original}{Keep the original file with the attribute/column names not changed?}

\item{verbosity}{Info?}
}
\value{
LMD data as data frame (if \code{spatial=FALSE}) or spatial points data frame (if \code{spatial=TRUE}) (invisible)
}
\description{
Download and/or load LUCAS Micro Data (LMD)
}

