% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmd_overview.R
\name{get_lmd_overview}
\alias{get_lmd_overview}
\title{Get the availability of LMD data by country and year}
\usage{
get_lmd_overview()
}
\value{
data frame
}
\description{
Get the availability of LMD data by country and year
}

