% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reclassify_lut.R
\name{get_reclassify_lut}
\alias{get_reclassify_lut}
\title{Build-in look-up tables for reclassifying LUCAS classes}
\usage{
get_reclassify_lut(type = "LC1nC9")
}
\arguments{
\item{type}{LC1nC9 or LC1nC10}
}
\value{
data.frame with the look up tabe that can be passed to \code{\link{reclassify_LULC}}.
}
\description{
Build-in look-up tables for reclassifying LUCAS classes
}
\examples{
\dontrun{
get_reclassify_lut("LC1nC9")
}
}

