% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_LMD_table.R
\name{info_LMD_table}
\alias{info_LMD_table}
\title{Print the description of a LUCAS attribute}
\usage{
info_LMD_table(attr = NULL, year = 2012, print = T)
}
\arguments{
\item{attr}{the attribute (see \code{\link{info_LMD_table_attributes}})}

\item{year}{the year of the LMD (so far only 2012 - attributes of other years migh differ)}

\item{print}{print the description? defaults to \code{TRUE}}
}
\value{
the information is returned invisible
}
\description{
Print the description of a LUCAS attribute
}

