% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_LMD_table.R
\name{info_LMD_table_attributes}
\alias{info_LMD_table_attributes}
\title{List the LMD attributes}
\usage{
info_LMD_table_attributes(year = 2012)
}
\arguments{
\item{year}{the year of the LMD (so far only 2012 - attributes of other years migh differ)}
}
\description{
List the LMD attributes
}

