% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas-internal.R
\docType{package}
\name{lucas}
\alias{lucas}
\alias{lucas-package}
\title{Utilities for Land cover/use statistics (LUCAS) data provided by EUROSTAT}
\description{
This package can be used to download the LUCAS Micro Data (LMD) and convert it into a spatial data set.
}

