% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Merge_harmo}
\alias{Merge_harmo}
\title{Merge all tables}
\usage{
Merge_harmo(con, rd)
}
\arguments{
\item{con}{Connection to db}

\item{rd}{Record descriptor in CSV format}
}
\description{
Merge all tables into a single harmonized version containing all years and change to relevant data type, as mapped in the record descriptor
}
\examples{
\dontrun{
Merge_harmo(con, '/data/LUCAS_harmo/data/supportDocs/LUCAS_harmo_RD.csv')}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
