% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_latlon_by_meter.R
\name{offset_latlon_by_meter}
\alias{offset_latlon_by_meter}
\title{Offset a latitude/longitude by some amount of meters}
\usage{
offset_latlon_by_meter(lat, lon, de = 100, dn = 100)
}
\arguments{
\item{lat}{latitude of the starting point(s) in decimal degree}

\item{lon}{longitude of the starting points(s) in decimal degree}

\item{de}{offset towards east (in meter)}

\item{dn}{offset towards north (in meter)}
}
\value{
latitud and longitude of the offsetted points
}
\description{
Offset a latitude/longitude by some amount of meters
}
\seealso{
http://gis.stackexchange.com/questions/2951/algorithm-for-offsetting-a-latitude-longitude-by-some-amount-of-meters
}

