% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Order_cols}
\alias{Order_cols}
\title{Change column order}
\usage{
Order_cols(con, years)
}
\arguments{
\item{con}{Connection to db}

\item{years}{Numeric vector of years to be harmonised}
}
\description{
Changes order of columns to fit the last survey (2018) and set all column data type to character varying in order to prepare for merge
}
\examples{
\dontrun{
Order_cols(con, c(2006, 2009, 2012, 2015))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
