% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_LMD_table.R
\name{print_attribute_info}
\alias{print_attribute_info}
\title{Print the description of a LMD attribute}
\usage{
print_attribute_info(attr = NULL, year = 2012)
}
\arguments{
\item{attr}{the attribute name (see \code{\link{info_LMD_table_attributes}})}

\item{year}{the year of the LMD (so far only 2012 - attributes of other years migh differ)}
}
\value{
the description (invisible)
}
\description{
Print the description of a LMD attribute
}

