% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify.R
\name{reclassify_LULC}
\alias{reclassify_LULC}
\title{Reclassifiy LUCAS data with a look-up-table}
\usage{
reclassify_LULC(x, lut)
}
\arguments{
\item{x}{the vector of classes to be reclassified}

\item{lut}{the look up table with}
}
\value{
vector with new classes
}
\description{
Reclassifiy LUCAS data with a look-up-table
}

