% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Remove_vars}
\alias{Remove_vars}
\title{Remove unwanted columns}
\usage{
Remove_vars(con, vars)
}
\arguments{
\item{con}{Connection to db}

\item{vars}{Character vector of variables, specified by name to remove}
}
\description{
Removes unwanted columns as specified by user
}
\examples{
\dontrun{
Remove_vars(con, vars)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
