% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{UF_Consistency_check}
\alias{UF_Consistency_check}
\title{User-friendly consistency check}
\usage{
UF_Consistency_check(con)
}
\arguments{
\item{con}{Connection to db}
}
\description{
Perform consistency checks on newly created UF fields to ensure conformity in terms of column order and data types
}
\examples{
\dontrun{
UF_Consistency_check(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
