% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Update_rd}
\alias{Update_rd}
\title{Update Record descriptor}
\usage{
Update_rd(con, rd, years)
}
\arguments{
\item{con}{Connection to db}

\item{rd}{Path to record descriptor csv}

\item{years}{Character vector of the years of survey}
}
\description{
Updates Record descriptor by adding a field (year) showing the year for which the variable exists and removing variables listed in Remove_vars function from RD
}
\examples{
\dontrun{
Update_rd(con, rd, years)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
