% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Upload_exif}
\alias{Upload_exif}
\title{Upload_exif}
\usage{
Upload_exif(con, exif)
}
\arguments{
\item{con}{Connection to database}

\item{exif}{the actual EXIF cvs located in mappings}
}
\description{
Upload_exif
}
\section{Functions}{
\itemize{
\item \code{Upload_exif}: to DB the exif information of LUCAS
}}

\examples{
\dontrun{
Add_photo_fields_2006n(con)}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
