% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{Upper_case}
\alias{Upper_case}
\title{Upper case columns}
\usage{
Upper_case(con, years)
}
\arguments{
\item{con}{Connection to db}

\item{years}{Numeric vector of years to be harmonised}
}
\description{
Convert values in designated columns (lc1, lc1_spec, lu1, lu1_type, lc2, lc2_spec, lu2, lu2_type, cprn_lc) to uppercase for consistency's sake
}
\examples{
\dontrun{
Upper_case(con,c(2009, 2012, 2015, 2018))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
