% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucas.r
\name{User_friendly}
\alias{User_friendly}
\title{User-friendly LUCAS harmonized}
\usage{
User_friendly(con, data_dir, years)
}
\arguments{
\item{con}{Connection to db}

\item{data_dir}{Directory where files are stored}

\item{years}{Numeric vector of years to be harmonised}
}
\description{
Creates columns with labels for coded variables and decodes all variables where possible to explicit labels
}
\examples{
\dontrun{
User_friendly(con, '/data/LUCAS_harmo/data/mappings', c(2006, 2009, 2012, 2015, 2018))}
}
\seealso{
To create the conection please see \link[lucas]{Connect_to_db}
}
